# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.SavedFreelancerResponse:
    properties:
      created_at:
        type: string
      freelancer_id:
        type: integer
      full_name:
        type: string
      id:
        type: integer
    type: object
  dto.SavedJobResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      job_id:
        type: integer
      job_title:
        type: string
    type: object
...
  utils.ErrorResponseSwagger:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API untuk manajemen jobseeker
  title: Jobseeker Management API
  version: "1.0"
paths:
...
  /saved/freelancers:
    get:
      consumes:
      - application/json
      description: Get list of freelancers saved by the company. Only companies can
        perform this action.
      produces:
      - application/json
      responses:
        "200":
          description: Freelancers retrieved successfully
          schema:
            items:
              $ref: '#/definitions/dto.SavedFreelancerResponse'
            type: array
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
        "403":
          description: 'Forbidden: Only companies can view saved freelancers'
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
        "500":
          description: Failed to retrieve freelancers
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
      security:
      - BearerAuth: []
      summary: Get Saved Freelancers
      tags:
      - saved
  /saved/freelancers/{freelancer_id}:
    delete:
      consumes:
      - application/json
      description: Remove a freelancer from the company's saved freelancers. Only
        companies can perform this action.
      parameters:
      - description: Freelancer ID
        in: path
        name: freelancer_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Freelancer removed from saved list successfully
          schema:
            $ref: '#/definitions/dto.SavedFreelancerResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
        "403":
          description: 'Forbidden: Only companies can remove saved freelancers'
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
        "500":
          description: Failed to remove saved freelancer
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
      security:
      - BearerAuth: []
      summary: Remove Saved Freelancer
      tags:
      - saved
  /saved/jobs:
    get:
      consumes:
      - application/json
      description: Get list of jobs saved by the freelancer. Only freelancers can
        perform this action.
      produces:
      - application/json
      responses:
        "200":
          description: Jobs retrieved successfully
          schema:
            items:
              $ref: '#/definitions/dto.SavedJobResponse'
            type: array
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
        "403":
          description: 'Forbidden: Only freelancers can view saved jobs'
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
        "500":
          description: Failed to retrieve jobs
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
      security:
      - BearerAuth: []
      summary: Get Saved Jobs
      tags:
      - saved
  /saved/jobs/{job_id}:
    delete:
      consumes:
      - application/json
      description: Remove a job from the freelancer's saved jobs. Only freelancers
        can perform this action.
      parameters:
      - description: Job ID
        in: path
        name: job_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Job removed from saved list successfully
          schema:
            $ref: '#/definitions/dto.SavedJobResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
        "403":
          description: 'Forbidden: Only freelancers can remove saved jobs'
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
        "500":
          description: Failed to remove saved job
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
      security:
      - BearerAuth: []
      summary: Remove Saved Job
      tags:
      - saved
    post:
      consumes:
      - application/json
      description: Save a job to the freelancer's saved jobs. Only freelancers can
        perform this action.
      parameters:
      - description: Job ID
        in: path
        name: job_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Job saved successfully
          schema:
            $ref: '#/definitions/dto.SavedJobResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
        "403":
          description: 'Forbidden: Only freelancers can save jobs'
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
        "500":
          description: Failed to save job
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
      security:
      - BearerAuth: []
      summary: Save Job
      tags:
      - saved
...
securityDefinitions:
  BearerAuth:
    description: Masukkan token dalam format "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"